Did lots of stuff
Sorry this is not a great commit.
Added new form element that dynamically creates
new buttons and values in a todoListName column
on Airtable.
Added logic when clicking the nav buttons it takes
you to the new list.
Auto populates the H1 based on the url using the
useParams() hook.